Separate validation functions from save signals
This will let us re-use validation functions for serializer validation
and for when a model saves. However, as of this commit, neither models'
save methods nor serializers' validation hook into the validators. The
hooks should be created soon.

In addition, for the task models which have arguments, conversion from
null to [] or {} needs to be done in separately (probably in a models
`clean()` function), since having these in the validators prevents them
from being idempotent, which a validator should be.