Rename tasks app to tasksapi
This requires either modifying or deleting all previous tasks
migrations. Deleting and remigrating is easier.

Tasks was a confusing name, since it's a common name of a
directory/module to hold Celery tasks, and tasksapi doesn't actually
hold any tasks in the normal sense—it just provides infrastructure for
them.