Disallow arbitrary dir bindings in task type
This shouldn't really be set in the task type. What a directory in a
container binds to on a host machine really depends on the host machine.
Given that task types should be host machine agnostic, this setup was no
good.

Instead we introduce a "results" directory that collects all non-log
output files, which the task type defines in the container, and the host
machine defines where to bind to in an environment variable.