Avoid nesting JSON for PK related fields
Given that the pagination stuff is in the first outer layer, then in the
next layer we have the results, nesting JSON for PK related fields
results in 3 layers of nesting, which results in too much non-flat
stuff. Better just to give the PK and make it flat (plus, this seems to
be what a few APIs I've looked at have opted for).