Name pulled Singularity containers by commit
These are for images pulled from Singularity Hub. This lets us keep many
different versions of a given image, which isn't necessarily taken
advantage of right now in the code, but hopefully will be in the near
future. (I guess there's nothing stopping anyone from taking advantage
of it now? Need to think about this one.)