Use a custom user model
I'm not sure if anyone is actually using saltant besides the Shah Lab,
but *if* you are what I did here is a painful procedure. What I did was
follow the steps in this thread

https://code.djangoproject.com/ticket/25313

and then after doing so I wiped migrations and made new ones. Please
email me at mwiens91@gmail.com if any of that is unclear.