Refactor tasksapi tests package
This commit does one major thing and one minor thing. The minor thing
is that it sequesters related test modules into sub-packages. The major
thing is that it provides a database fixture that test cases can use,
which obviates the need for having explicit tests for model creation
(since since using the provided fixture in any test automatically
creates the included models).